modularized the structure of backend code
