+++ backend/__init__.py
@@ -1,22 +1,30 @@
 from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from flask_marshmallow import Marshmallow
-from flask_migrate import Migrate
 from .config import config
 from flask_cors import CORS
 from dotenv import load_dotenv
-# Initialize extensions
-db = SQLAlchemy()
-ma = Marshmallow()
-migrate = Migrate()
+from .extensions import db, ma, migrate
+from .routes.accommodation_routes import accommodation_routes
+from .routes.document_routes import document_routes
 
 def create_app(config_mode):
+    """
+    Create a Flask application instance with a specified configuration.
+
+    This function sets up the application, including loading environment variables,
+    initializing extensions such as SQLAlchemy, Marshmallow, and Flask-Migrate, 
+    and registering the necessary blueprints (routes).
+
+    :param config_mode: The configuration mode (e.g., 'development', 'production') as specified in the config.py file.
+    :return: Configured Flask application instance.
+    """
+    # Load environment variables
     load_dotenv()
 
     # Create Flask application
     app = Flask(__name__)
     CORS(app)
-    # Set config
+
+    # Load app configuration based on the mode (development/production)
     app.config.from_object(config[config_mode])
 
     # Initialize extensions with app
@@ -24,8 +32,8 @@ def create_app(config_mode):
     ma.init_app(app)
     migrate.init_app(app, db)
 
-    # Import and register blueprints
-    from .routes import accommodation_routes
+    # Register blueprints for accommodation and document routes
     app.register_blueprint(accommodation_routes)
+    app.register_blueprint(document_routes)
 
     return app
