+++ backend/extensions.py
@@ -1,0 +1,8 @@
+# extensions.py
+from flask_sqlalchemy import SQLAlchemy
+from flask_marshmallow import Marshmallow
+from flask_migrate import Migrate
+
+db = SQLAlchemy()
+ma = Marshmallow()
+migrate = Migrate()
